restructuring for pytest
