+++ tests/test_mood_ring.py
@@ -1,4 +1,6 @@
-from context import mood_ring as mr
+#import sys
+#sys.path.append('..')
+import mood_ring as mr
 
 def test_ring_output_membership():
     mood_list = [item[0] for item in mr.mood_map]
